Alterado classe principal
